backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
aiofiles
PyYAML
python-dotenv
black
pickledb
